+++ src/app/ExpressServerAddOn.ts
@@ -1,3 +1,6 @@
+/// <reference types="debug" />
+const debug: debug.IDebugger = require('debug')('@micro-fleet/web:ExpressServerAddOn')
+
 import * as fs from 'fs'
 import * as path from 'path'
 import * as http from 'http'
@@ -310,7 +313,7 @@ export class ExpressServerAddOn implements IServiceAddOn {
             this._server = server
                 .on('listening', () => {
                     this._isAlive = true
-                    console.log('HTTP listening on: ', this._port)
+                    debug('HTTP listening on: %d', this._port)
                     resolve()
                 })
                 .on('error', reject)
@@ -329,7 +332,7 @@ export class ExpressServerAddOn implements IServiceAddOn {
             this._sslServer = https.createServer(sslOptions, app)
                 .on('listening', () => {
                     this._isAlive = true
-                    console.log('HTTPS listening on: ', this._sslPort)
+                    debug('HTTPS listening on: %d', this._sslPort)
                     resolve()
                 })
                 .on('error', reject)
@@ -414,14 +417,12 @@ export class ExpressServerAddOn implements IServiceAddOn {
         // Otherwise, a new controller instance will be created for each request.
         return HandlerContainer.instance.register(actionFunc.name, CtrlClass.name,
             (ctrlInstance, actionName) => {
-                if (actionName !== actionFunc.name) { return null }
-
                 // Wrapper function that handles uncaught errors,
                 // so that controller actions don't need to call `next(error)` like said
                 // by https://expressjs.com/en/guide/error-handling.html
                 return function (this: any, req: Request, res: Response, next: express.NextFunction) {
                     try {
-                        const call = actionFunc.call(this, req, res)
+                        const call = ctrlInstance[actionName](req, res)
                         // Catch async exception
                         if (call && typeof call.catch === 'function') {
                             call.catch(next)